Move method 'toDateString' from DateTimeUtil to DateTimeFormatterUtil
