+++ game-core/src/main/java/games/strategy/engine/chat/ChatMessagePanel.java
@@ -26,8 +26,8 @@ import javax.swing.text.SimpleAttributeSet;
 import javax.swing.text.StyleConstants;
 import lombok.extern.java.Log;
 import org.triplea.domain.data.UserName;
-import org.triplea.java.Interruptibles;
 import org.triplea.java.DateTimeUtil;
+import org.triplea.java.Interruptibles;
 import org.triplea.sound.ClipPlayer;
 import org.triplea.sound.SoundPath;
 import org.triplea.swing.SwingAction;
