+++ java-extras/src/test/java/org/triplea/java/DateTimeFormatterUtilTest.java
@@ -11,6 +11,9 @@ import org.junit.jupiter.api.Test;
 import org.triplea.java.DateTimeFormatterUtil.FormatOption;
 
 class DateTimeFormatterUtilTest {
+  private static final Instant JAN_FIRST_INSTANT =
+      LocalDateTime.of(2020, 1, 1, 14, 30).toInstant(ZoneOffset.UTC);
+
   private static final long DEC_FIRST_EPOCH_MILLIS =
       LocalDateTime.of(2000, 12, 1, 23, 59) //
           .toInstant(ZoneOffset.ofHours(8))
@@ -34,5 +37,13 @@ class DateTimeFormatterUtilTest {
         DateTimeFormatterUtil.formatEpochMilli(
             DEC_FIRST_EPOCH_MILLIS, FormatOption.WITHOUT_TIMEZONE);
     assertThat(result, is("2000-12-1 15:59"));
+  }
+
+  @Test
+  void toDateString() {
+    assertThat(
+        DateTimeFormatterUtil.toDateString(
+            LocalDateTime.ofInstant(JAN_FIRST_INSTANT, ZoneOffset.UTC)),
+        is("Wed. Jan. 01 14:30:00 Z 2020"));
   }
 }
